refactor: datetime UTC fix, design tokens, price float, interview fallback, token expiry, exception fix
datetime.utcnow() → datetime.now(UTC).strftime():
- Replace all 13 datetime.utcnow() calls with timezone-aware datetime.now(UTC)
- Use strftime instead of isoformat to avoid +00:00Z invalid format
- All DeprecationWarnings eliminated (68 → 0)

CSS design tokens:
- Add spacing scale (--sp-xs through --sp-2xl)
- Add typography tokens (--font-sm through --font-3xl, --weight-*)
- Add --radius-full, --shadow-sm

price_cny: int → float:
- schemas.py PricingPackage, PaymentOrderResponse, PaymentCreateRequest

interview questions fallback:
- Add _fallback_questions() with context-specific questions based on gaps
- Uses gap_type to generate skill_gap/evidence_missing/unknown questions

token expiration (90 days):
- users table: add last_used_at column via migration
- get_user_by_token: checks last_used_at, updates on each use
- register_user: sets last_used_at on creation
- Fix except (ValueError, Exception) catching HTTPException (re-raise)

tests:
- test_get_user_by_token_expired: verifies 91-day-old token rejected
- test_get_user_by_token_updates_last_used: verifies last_used_at set

Pytest: 51 passed, 0 warnings

+++ app/schemas.py
@@ -101,7 +101,7 @@ class ProviderCatalogResponse(BaseModel):
 class PricingPackage(BaseModel):
     code: str
     name: str
-    price_cny: int
+    price_cny: float
     credits: int
     description: str
     includes: List[str]
@@ -139,7 +139,7 @@ class PurchaseResponse(BaseModel):
     package_name: str
     credits_added: int
     credits_total: int
-    price_cny: int
+    price_cny: float
 
 
 class PaymentOrderResponse(BaseModel):
@@ -147,7 +147,7 @@ class PaymentOrderResponse(BaseModel):
     created_at: str
     package_name: str
     credits: int
-    price_cny: int
+    price_cny: float
     status: str
     payment_method: str
